Change toolbar in CurrencySwitcherFragment
