+++ components/currencyswitcher/build.gradle
@@ -35,11 +35,11 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.3.0-alpha01'
-    implementation 'androidx.core:core-ktx:1.3.0'
+    implementation 'androidx.appcompat:appcompat:1.3.0-alpha02'
+    implementation 'androidx.core:core-ktx:1.3.2'
 
-    implementation 'com.google.android.material:material:1.3.0-alpha01'
-    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'com.google.android.material:material:1.3.0-alpha03'
+    implementation 'androidx.recyclerview:recyclerview:1.2.0-alpha06'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     implementation project(':core')
