+++ components/currencyswitcher/src/main/res/layout/view_holder_currency_switcher_footer.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    style="@style/Subhead2"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:paddingStart="24dp"
+    android:paddingTop="12dp"
+    android:paddingEnd="24dp"
+    android:paddingBottom="12dp"
+    android:text="@string/SettingsCurrency_InfoText"
+    android:textColor="@color/grey" />
