+++ core/build.gradle
@@ -31,18 +31,18 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.3.0-alpha01'
+    implementation 'androidx.appcompat:appcompat:1.3.0-alpha02'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    implementation 'androidx.core:core-ktx:1.3.0'
+    implementation 'androidx.core:core-ktx:1.3.2'
 
     implementation 'androidx.fragment:fragment-ktx:1.3.0-beta01'
     // Navigation component
-    implementation "androidx.navigation:navigation-fragment-ktx:2.3.0"
-    implementation "androidx.navigation:navigation-ui-ktx:2.3.0"
+    api "androidx.navigation:navigation-fragment-ktx:2.3.1"
+    api "androidx.navigation:navigation-ui-ktx:2.3.1"
 
     implementation 'io.reactivex.rxjava2:rxjava:2.2.19'
     implementation "androidx.biometric:biometric:1.0.1"
-    implementation 'com.google.android.material:material:1.3.0-alpha01'
+    implementation 'com.google.android.material:material:1.3.0-alpha03'
     implementation project(':components:views')
     implementation project(':components:snackbar')
